[#9] 코드 정리, 이름 변경, 주석 추가 - CompanyNewsListPage 마크업 개발 준비 작업 - 주석 추가 - NewsCard 컴포넌트를 News 로 변경하고 templates 에서 components 풀더로 변경 - RecommendNewsCard 컴포넌트 이름을 RecommendNewsList 로 변경
