+++ src/components/index.ts
@@ -1,1 +1,2 @@
 export * from './ExternalLink';
+export * from './News';
