+++ src/pages/index.ts
@@ -2,3 +2,4 @@ export { default as CompanySelectPage } from './CompanySelectPage';
 export { default as HomePage } from './HomePage';
 export { default as PicksPage } from './PicksPage';
 export { default as NewsPage } from './NewsPage';
+export { default as CompanyNewsListPage } from './CompanyNewsListPage';
