+++ src/templates/NewsList/RecommendNewsList.tsx
@@ -1,0 +1,52 @@
+import React from 'react';
+import styled from 'styled-components';
+
+import { ColListLayout, RowListLayout } from 'Layouts/index';
+import { getRecommendNewsList } from 'Api/index';
+
+import CardSimpleLayout from './CardSimpleLayout';
+
+const RecommendNewsList = () => {
+  const recommendNewsList = getRecommendNewsList();
+
+  return (
+    <CardSimpleLayout>
+      {{
+        header: <RecommendNewsListTitle />,
+        body: (
+          <ColListLayout.Repeat interval="15px">
+            {recommendNewsList.map(({ link, title, key }) => (
+              <RowListLayout.Repeat interval="10px" key={key}>
+                <RecommendNewsThumbnail src={link.img} />
+                <RecommendNewsTitle>{title}</RecommendNewsTitle>
+              </RowListLayout.Repeat>
+            ))}
+          </ColListLayout.Repeat>
+        ),
+      }}
+    </CardSimpleLayout>
+  );
+};
+
+const RecommendNewsListTitle = styled.div.attrs({
+  children: '커머스, 이런 기사도 읽어보세요.',
+})`
+  font-family: SFProDisplay;
+  font-size: 17px;
+  font-weight: 500;
+  color: #121111;
+`;
+
+const RecommendNewsThumbnail = styled.img`
+  width: 60px;
+  height: 60px;
+`;
+
+const RecommendNewsTitle = styled.div`
+  font-family: SFProDisplay;
+  font-size: 16px;
+  line-height: 1.25;
+  color: #121111;
+`;
+
+export default RecommendNewsList;
