+++ src/templates/Readme.md
@@ -1,0 +1,3 @@
+비지니스 로직이 있는 컴포넌트 (컨테이너 컴포넌트)
+
+mobx 적용
