+++ src/templates/index.ts
@@ -9,4 +9,3 @@ export { default as Tabs } from './Tabs';
 export { default as PickNewsList } from './PickNewsList';
 export { default as PickCategoryList } from './PickCategoryList';
 export { default as TextCard } from './TextCard';
-export { default as NewsCard } from './NewsCard';
